# Conflicts:
#	compiler/assembler.cpp
#	compiler/assembler.hpp
#	compiler/ast/ast.cpp
#	compiler/ast/ast.hpp
#	compiler/ast/ast_printer.cpp
#	compiler/compiler.hpp
#	compiler/hir/type_checker.cpp
#	compiler/mir/mir.cpp
#	compiler/mir/mir.hpp
#	compiler/parser.cpp
#	lmx.cpp
#	runtime/CMakeLists.txt
#	runtime/binary.hpp
#	runtime/object/code_module.cpp
#	runtime/object/object.cpp
#	runtime/object/object.hpp
#	runtime/object/value.cpp
#	runtime/object/value.hpp
#	runtime/opcode.hpp
#	runtime/vm.cpp
#	utils/utils.hpp

+++ runtime/object/fraction.cpp
@@ -166,7 +166,7 @@ bool Fraction::operator!=(const Fraction& other) const noexcept {
 }
 
 double Fraction::to_float() const noexcept {
-    return static_cast<double>(num) / den;
+    return static_cast<double>(num) / static_cast<double>(den);
 }
 
 bool Fraction::operator!=(const int64_t other) const noexcept {
